requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyodbc==5.0.1
google-generativeai==0.3.1
schedule==1.2.0
pydantic==2.5.0
python-multipart==0.0.6